Added `iota` mode, and fixed a bug with Extract mode.
The bug was where extract was using an `if val` test
to detect whether a value was successfully extracted..
but obviously this was dumb, because it'll fail to
print falsey values like zero. So, now it checks
`if val is not None` which still isn't perfect,
but is probably useful enough for now.

`iota` mode meanwhile scratches a little itch where
one might want to process a large file and make a
small edit or addition to every object. It provides
an iterating mode (with optional in-stream deduping)
where arbitrary objectpath queries are assigned to
a key, which can overwrite existing keys. As a
special addition, a key `__iota` is added to each
object which contains the zero-indexed index of
the object in the file. This value does not
increment when deduping skips a line. So a simple
case (which the mode was written for, actually)
is `jltool iota 'id' '$.__iota' file.jl > indexed.jl`,
which simply adds an `id` key to each object with
the integer index of the object in the file. Simple!